backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=1.0.0
psycopg2-binary>=2.9
python-dotenv>=1.0
